fix: overflow in ProgressStyle::named_bar

+++ src/util/named_progress.rs
@@ -7,7 +7,7 @@ pub trait NamedProgress {
 impl NamedProgress for ProgressStyle {
     fn named_bar(name: &str) -> Self {
         let mut fmt = name.to_string();
-        for _ in 0..(10 - name.len() - 1) {
+        for _ in 0..(10 - name.len() as i64 - 1) {
             fmt += " ";
         }
         fmt += " {bar:40.cyan/blue} {pos:>6}/{len:<6} [{elapsed_precise} ({eta} remaining)] {msg}";
